connect main/app to store
