+++ package.json
@@ -11,8 +11,10 @@
     "ngrok": "^3.2.7",
     "react": "16.8.3",
     "react-native": "0.59.9",
+    "react-redux": "^7.2.1",
     "react-timer-mixin": "^0.13.4",
-    "react-viro": "2.17.0"
+    "react-viro": "2.17.0",
+    "redux": "^4.0.5"
   },
   "devDependencies": {
     "@babel/core": "^7.11.1",
